pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "render_engine"
dynamic = ["version"]
description = "A Flexible Static Site Generator for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "textualize",
  "textual-web",
  "render-engine",
]

[project.optional-dependencies]
dev = [
    "ruff",
    ]

[tool.setuptools_scm]
local_scheme = "no-local-version"
# version_scheme = "python-simplified-semver"

[project.urls]
homepage = "https://github.com/kjaymiller/render_engine/"
repository = "https://github.com/kjaymiller/render_engine/"
documentation = "https://render-engine.readthedocs.io/en/latest/"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.semantic_release]
version_toml = "pyproject.toml:project.version"
branch = "main"

[too.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

